+++ Models/Public/ResourceSchemaType.cs
@@ -1,0 +1,9 @@
+namespace EINVWORLD.Models.Public
+{
+    public enum ResourceSchemaType
+    {
+        Article = 0,
+        FAQ = 1,
+        HowTo = 2
+    }
+}